+++ src/main/java/com/caffeine/gwanghwamun/domain/cart/entity/CartItemOption.java
@@ -1,0 +1,4 @@
+package com.caffeine.gwanghwamun.domain.cart.entity;
+
+public class CartItemOption {
+}
