+++ src/main/java/com/caffeine/gwanghwamun/domain/order/dto/OrderItemOptionId.java
@@ -1,0 +1,6 @@
+package com.caffeine.gwanghwamun.domain.order.dto;
+
+import java.util.UUID;
+
+public record OrderItemOptionId(UUID id) {
+}
