+++ src/main/java/com/caffeine/gwanghwamun/domain/order/dto/OrderItemOptionResDTO.java
@@ -1,0 +1,7 @@
+package com.caffeine.gwanghwamun.domain.order.dto;
+
+import java.util.UUID;
+
+public record OrderItemOptionResDTO(
+    UUID orderItemOptionId, String orderItemOptionName, Integer price) {
+}
