--- src/main/java/com/caffeine/gwanghwamun/domain/order/order_items/dto/OrderItemOptionResDTO.java
@@ -1,6 +1,0 @@
-package com.caffeine.gwanghwamun.domain.order.order_items.dto;
-
-import java.util.UUID;
-
-public record OrderItemOptionResDTO(
-		UUID orderItemOptionId, String orderItemOptionName, Integer price) {}
