--- src/main/java/com/caffeine/gwanghwamun/domain/order/order_items/order_item_options/dto/OrderItemOptionId.java
@@ -1,5 +1,0 @@
-package com.caffeine.gwanghwamun.domain.order.order_items.order_item_options.dto;
-
-import java.util.UUID;
-
-public record OrderItemOptionId(UUID id) {}
