--- src/main/java/com/caffeine/gwanghwamun/domain/order/order_items/repository/OrderItemRepository.java
@@ -1,10 +1,0 @@
-package com.caffeine.gwanghwamun.domain.order.order_items.repository;
-
-import com.caffeine.gwanghwamun.domain.order.order_items.entity.OrderItem;
-import java.util.List;
-import java.util.UUID;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface OrderItemRepository extends JpaRepository<OrderItem, UUID> {
-	List<OrderItem> findByOrder_OrderIdAndDeletedDateIsNull(UUID orderId);
-}
