--- src/main/java/com/caffeine/gwanghwamun/domain/order/order_status_log/dto/OrderStatusLogResDTO.java
@@ -1,28 +1,0 @@
-package com.caffeine.gwanghwamun.domain.order.order_status_log.dto;
-
-import com.caffeine.gwanghwamun.domain.order.entity.OrderStatus;
-import com.caffeine.gwanghwamun.domain.order.order_status_log.entity.OrderStatusLog;
-import java.time.LocalDateTime;
-import java.util.UUID;
-
-public record OrderStatusLogResDTO(
-		UUID orderStatusLogId,
-		UUID orderId,
-		Long userId,
-		UUID storeId,
-		OrderStatus existingState,
-		OrderStatus currentState,
-		String reason,
-		LocalDateTime createAt) {
-	public OrderStatusLogResDTO(OrderStatusLog orderStatusLog) {
-		this(
-				orderStatusLog.getOrderStatusLogId(),
-				orderStatusLog.getOrder().getOrderId(),
-				orderStatusLog.getUser().getUserId(),
-				orderStatusLog.getStore().getStoreId(),
-				orderStatusLog.getExistingState(),
-				orderStatusLog.getCurrentState(),
-				orderStatusLog.getReason(),
-				orderStatusLog.getCreateAt());
-	}
-}
