+++ src/main/java/com/caffeine/gwanghwamun/domain/order/repository/OrderItemOptionRepository.java
@@ -1,0 +1,10 @@
+package com.caffeine.gwanghwamun.domain.order.repository;
+
+import com.caffeine.gwanghwamun.domain.order.entity.OrderItemOption;
+
+import java.util.UUID;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface OrderItemOptionRepository extends JpaRepository<OrderItemOption, UUID> {
+}
